refactor(operations): TASK-028 extract WorkdayPanel (EPIC-006 phase 0)
Make the field-workday UI an importable component so the Technician My Day
surface can reuse it (EPIC-006). WorkdayPanel.tsx is the former DailyCommandCenter
verbatim (only the function is renamed); DailyCommandCenter.tsx is now a thin
alias re-exporting it, so /app and all importers are unchanged. Zero behavior
change. The field/business separation (action queue, revenue widgets embedded in
the field tabs) is deferred to phase 2 (TASK-030).

Verified: typecheck, lint, and production build pass; /app compiles.
